split out operations, handlers, and internals
