+++ chirho/dynamical/ODE/__init__.py
@@ -8,15 +8,11 @@ import pyro
 import torch
 import torchdiffeq
 
-from chirho.dynamical.internals import (
-    State,
-    Trajectory,
-    apply_interruptions,
+from chirho.dynamical.internals.interruption import (
     concatenate,
-    simulate_to_interruption,
 )
-
-from chirho.dynamical.ops import Dynamics, simulate
+from chirho.dynamical.internals.interruption import apply_interruptions, simulate_to_interruption
+from chirho.dynamical.ops import Dynamics, simulate, State, Trajectory
 from chirho.indexed.ops import IndexSet, gather, indices_of, union
 from chirho.interventional.handlers import intervene
 from chirho.observational.handlers import condition
