+++ chirho/dynamical/handlers.py
@@ -8,67 +8,20 @@ import pyro
 import torch
 import torchdiffeq
 
-from chirho.dynamical.internals import (
-    State,
-    Trajectory,
+from chirho.dynamical.ops import State, Trajectory
+from chirho.dynamical.internals.indexed import gather, indices_of
+
+# from chirho.indexed.ops import IndexSet, gather, indices_of, union
+from chirho.dynamical.internals.interruption import (
     apply_interruptions,
-    concatenate,
     simulate_to_interruption,
+    concatenate
 )
-from chirho.dynamical.ops import simulate
-from chirho.indexed.ops import IndexSet, gather, indices_of, union
-from chirho.interventional.handlers import intervene
+from chirho.dynamical.internals.interventional import intervene
 from chirho.observational.handlers import condition
 
 S = TypeVar("S")
 T = TypeVar("T")
-
-
-@indices_of.register
-def _indices_of_state(state: State, *, event_dim: int = 0, **kwargs) -> IndexSet:
-    return union(
-        *(
-            indices_of(getattr(state, k), event_dim=event_dim, **kwargs)
-            for k in state.keys
-        )
-    )
-
-
-@indices_of.register
-def _indices_of_trajectory(
-    trj: Trajectory, *, event_dim: int = 0, **kwargs
-) -> IndexSet:
-    return union(
-        *(
-            indices_of(getattr(trj, k), event_dim=event_dim + 1, **kwargs)
-            for k in trj.keys
-        )
-    )
-
-
-@gather.register(State)
-def _gather_state(
-    state: State[T], indices: IndexSet, *, event_dim: int = 0, **kwargs
-) -> State[T]:
-    return type(state)(
-        **{
-            k: gather(getattr(state, k), indices, event_dim=event_dim, **kwargs)
-            for k in state.keys
-        }
-    )
-
-
-@gather.register(Trajectory)
-def _gather_trajectory(
-    trj: Trajectory[T], indices: IndexSet, *, event_dim: int = 0, **kwargs
-) -> Trajectory[T]:
-    return type(trj)(
-        **{
-            k: gather(getattr(trj, k), indices, event_dim=event_dim + 1, **kwargs)
-            for k in trj.keys
-        }
-    )
-
 
 
 class SimulatorEventLoop(Generic[T], pyro.poutine.messenger.Messenger):
@@ -230,16 +183,6 @@ class PointInterruption(Interruption):
         # we put it in the simulate preprocess call instead, to get the full timespan. elif self.time < start_time:
 
         super()._pyro_simulate_to_interruption(msg)
-
-
-@intervene.register(State)
-def state_intervene(obs: State[T], act: State[T], **kwargs) -> State[T]:
-    new_state: State[T] = State()
-    for k in obs.keys:
-        setattr(
-            new_state, k, intervene(getattr(obs, k), getattr(act, k, None), **kwargs)
-        )
-    return new_state
 
 
 class _InterventionMixin(Interruption):
