+++ tests/dynamical/test_dynamic_interventions.py
@@ -8,15 +8,9 @@ from chirho.counterfactual.handlers import (
     MultiWorldCounterfactual,
     TwinWorldCounterfactual,
 )
-from chirho.dynamical.handlers import (
-    DynamicIntervention,
-    SimulatorEventLoop,
-    simulate,
-)
-
+from chirho.dynamical.handlers import DynamicIntervention, SimulatorEventLoop
 from chirho.dynamical.ODE import ODEDynamics
-
-from chirho.dynamical.internals import State
+from chirho.dynamical.ops import simulate, State
 from chirho.indexed.ops import IndexSet, gather, indices_of, union
 
 from .dynamical_fixtures import UnifiedFixtureDynamics
